backend/llm/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
Booktype==1.5
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
distro==1.9.0
Django==5.1.6
fastapi==0.115.8
flake8==7.1.1
google==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mccabe==0.7.0
numpy==2.1.0
openai==1.63.0
packaging==24.2
paramiko==3.5.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
router==0.1
rsa==4.9
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.45.3
storage==0.0.4.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
redis~=5.2.1
requests~=2.32.3
PyJWT~=2.10.1
google-cloud-bigquery
google~=3.0.0
protobuf~=5.29.3
pydantic-settings~=2.8.1